requirements.txt
asgiref==3.5.2
beautifulsoup4==4.9.3
certifi==2022.6.15
chardet==4.0.0
DAWG-Python==0.7.2
Django==4.0.5
docopt==0.6.2
google==3.0.0
googlesearch-python==1.1.0
idna==2.10
pdfminer==20191125
pycryptodome==3.15.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
requests==2.25.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
yargy==0.15.0
